+++ src/gakudan_sup.erl
@@ -10,7 +10,7 @@ start_link() ->
     supervisor:start_link({local, ?MODULE}, ?MODULE, []).
 
 init([]) ->
-    SupFlags = #{strategy => one_for_one, intensity => 5, period => 10},
+    SupFlags = #{strategy => rest_for_one, intensity => 5, period => 10},
     Children = [
         #{
             id => gakudan_registry,
@@ -21,6 +21,12 @@ init([]) ->
             id => gakudan_runs_sup,
             start => {gakudan_runs_sup, start_link, []},
             type => supervisor
+        },
+        #{
+            id => gakudan_runs_resumer,
+            start => {gakudan_runs_resumer, start_link, []},
+            restart => transient,
+            type => worker
         }
     ],
     {ok, {SupFlags, Children}}.